Reforming Login, SignUp View Layout Designs
