+++ DeepFashion/Model/Constants.swift
