+++ DeepFashion/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
