+++ DeepFashion/SupportingFiles/Assets.xcassets/Contents.json
