+++ DeepFashion/ViewController/SignUp/FirstSignUpViewController.swift
@@ -1,0 +1,34 @@
+//
+//  SignUpViewController.swift
+//  DeepFashion
+//
+//  Created by MinKyeongTae on 27/10/2019.
+//  Copyright © 2019 MinKyeongTae. All rights reserved.
+//
+
+import UIKit
+import RxSwift
+import RxCocoa
+
+class FirstSignUpViewController: UIViewController {
+    
+    /// MARK: - IBOutlet
+    @IBOutlet private weak var idTextField: UITextField!
+    @IBOutlet private weak var passwordTextField: UITextField!
+    @IBOutlet private weak var passwordConfirmTextField: UITextField!
+    @IBOutlet private weak var genderSegmentedControl: UISegmentedControl!
+    @IBOutlet private weak var nextPageButton: UIButton!
+    
+    /// MARK: - Properties
+
+    
+    /// MARK: - LifeCycle
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+    }
+    
+    /// MARK: - Method
+
+    
+}
